+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - Feature: `CLIENT KILL` now support `LADDR` argument.
 - Feature: `CLIENT LIST` now support `ID` argument.
--
+- Feature: `CLIENT PAUSE` support options and added `CLIENT UNPAUSE` command.
 
 ## 1.11.1